+++ Tutorials/12_handwriting_recognition_2/configs.py
@@ -6,12 +6,12 @@ from mltu.configs import BaseModelConfigs
 class ModelConfigs(BaseModelConfigs):
     def __init__(self):
         super().__init__()
-        self.model_path = os.path.join("Models/03_handwriting_recognition", datetime.strftime(datetime.now(), "%Y%m%d%H%M"))
+        self.model_path = os.path.join("Models/12_handwriting_recognition_2", datetime.strftime(datetime.now(), "%Y%m%d%H%M"))
         self.vocab = ""
         self.height = 32
         self.width = 128
         self.max_text_length = 0
         self.batch_size = 128
-        self.learning_rate = 0.001
+        self.learning_rate = 0.002
         self.train_epochs = 1000
         self.train_workers = 20