+++ Tutorials/12_handwriting_recognition_2/inferenceModel.py
@@ -1,0 +1,48 @@
+import cv2
+import typing
+import numpy as np
+
+from mltu.inferenceModel import OnnxInferenceModel
+from mltu.utils.text_utils import ctc_decoder, get_cer
+
+class ImageToWordModel(OnnxInferenceModel):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def predict(self, image: np.ndarray):
+        image = cv2.resize(image, self.input_shapes[0][1:3][::-1])
+
+        image_pred = np.expand_dims(image, axis=0).astype(np.float32)
+
+        preds = self.model.run(self.output_names, {self.input_names[0]: image_pred})[0]
+
+        text = ctc_decoder(preds, self.model.metadata["vocab"])[0]
+
+        return text
+
+if __name__ == "__main__":
+    import pandas as pd
+    from tqdm import tqdm
+
+    model = ImageToWordModel(model_path="Models/12_handwriting_recognition_2/202503281916/model.onnx",)
+
+    df = pd.read_csv("Models/03_handwriting_recognition/202301111911/val.csv").values.tolist()
+
+    accum_cer = []
+    for image_path, label in tqdm(df):
+        image = cv2.imread(image_path.replace("\\", "/"))
+
+        prediction_text = model.predict(image)
+
+        cer = get_cer(prediction_text, label)
+        print(f"Image: {image_path}, Label: {label}, Prediction: {prediction_text}, CER: {cer}")
+
+        accum_cer.append(cer)
+
+        # resize by 4x
+        image = cv2.resize(image, (image.shape[1] * 4, image.shape[0] * 4))
+        cv2.imshow("Image", image)
+        cv2.waitKey(0)
+        cv2.destroyAllWindows()
+
+    print(f"Average CER: {np.average(accum_cer)}")