+++ Tutorials/12_handwriting_recognition_2/train.py
@@ -1,0 +1,111 @@
+import os
+#visible 1 gpu
+os.environ["CUDA_VISIBLE_DEVICES"] = "1"
+import tensorflow as tf
+try: [tf.config.experimental.set_memory_growth(gpu, True) for gpu in tf.config.experimental.list_physical_devices("GPU")]
+except: pass
+
+import keras
+
+from mltu.preprocessors import ImageReader
+from mltu.transformers import ImageResizer, LabelIndexer, LabelPadding, ImageShowCV2
+from mltu.augmentors import RandomBrightness, RandomRotate, RandomErodeDilate, RandomSharpen, RandomElasticTransform, RandomGaussianBlur, RandomSaltAndPepper
+from mltu.annotations.images import CVImage
+
+
+from mltu.tensorflow.dataProvider import DataProvider
+from mltu.tensorflow.losses import CTCloss
+from mltu.tensorflow.callbacks import Model2onnx, TrainLogger
+from mltu.tensorflow.metrics import CERMetric, WERMetric
+
+from model import train_model
+from configs import ModelConfigs
+
+import os
+import tarfile
+from data_utils import download_and_unzip, load_dataset
+
+dataset_path = os.path.join("Datasets", "IAM_Words")
+if not os.path.exists(dataset_path):
+    download_and_unzip("https://git.io/J0fjL", extract_to="Datasets")
+
+    file = tarfile.open(os.path.join(dataset_path, "words.tgz"))
+    file.extractall(os.path.join(dataset_path, "words"))
+
+dataset, vocab, max_len = load_dataset(dataset_path)
+
+# Create a ModelConfigs object to store model configurations
+configs = ModelConfigs()
+configs.vocab = "".join(sorted(vocab))
+configs.max_text_length = max_len
+configs.save()
+
+# Create a data provider for the dataset
+data_provider = DataProvider(
+    dataset=dataset,
+    skip_validation=True,
+    batch_size=configs.batch_size,
+    data_preprocessors=[ImageReader(CVImage)],
+    transformers=[
+        ImageResizer(configs.width, configs.height, keep_aspect_ratio=False),
+        LabelIndexer(configs.vocab),
+        LabelPadding(max_word_length=configs.max_text_length, padding_value=len(configs.vocab)),
+        ],
+    use_cache=True,
+    workers=configs.train_workers,
+    max_queue_size=10,
+    # use_multiprocessing=True
+)
+
+# Split the dataset into training and validation sets
+train_data_provider, val_data_provider = data_provider.split(split = 0.9)
+
+# Augment training data with random brightness, rotation and erode/dilate
+train_data_provider.augmentors = [
+    RandomBrightness(), 
+    RandomErodeDilate(),
+    RandomSharpen(),
+    RandomRotate(angle=10), 
+    RandomElasticTransform(),
+    RandomGaussianBlur(),
+    RandomSaltAndPepper(),
+]
+
+# Creating TensorFlow model architecture
+model = train_model(
+    input_dim = (configs.height, configs.width, 3),
+    output_dim = len(configs.vocab),
+)
+
+# Compile the model and print summary
+model.compile(
+    optimizer=keras.optimizers.Nadam(learning_rate=configs.learning_rate), # pyright: ignore
+    loss=CTCloss(), 
+    metrics=[
+        CERMetric(vocabulary=configs.vocab),
+        WERMetric(vocabulary=configs.vocab)
+    ],
+    jit_compile=False, # pyright: ignore
+    run_eagerly=False,
+)
+model.summary(line_length=110)
+
+# Define callbacks
+earlystopper = keras.callbacks.EarlyStopping(monitor="val_CER", patience=20, verbose=1, mode="min")
+checkpoint = keras.callbacks.ModelCheckpoint(f"{configs.model_path}/model.h5", monitor="val_CER", verbose=1, save_best_only=True, mode="min")
+trainLogger = TrainLogger(configs.model_path)
+tb_callback = keras.callbacks.TensorBoard(f"{configs.model_path}/logs", update_freq="epoch")
+reduceLROnPlat = keras.callbacks.ReduceLROnPlateau(monitor="val_CER", factor=0.9, min_delta=1e-10, patience=5, verbose=1, mode="min")
+model2onnx = Model2onnx(f"{configs.model_path}/model.h5", metadata={"vocab": configs.vocab}, save_on_epoch_end=True, opset=18)
+
+# Save training and validation datasets as csv files
+train_data_provider.to_csv(os.path.join(configs.model_path, "train.csv"))
+val_data_provider.to_csv(os.path.join(configs.model_path, "val.csv"))
+
+# Train the model
+model.fit(
+    train_data_provider,
+    validation_data=val_data_provider,
+    epochs=configs.train_epochs,
+    callbacks=[earlystopper, checkpoint, trainLogger, reduceLROnPlat, tb_callback, model2onnx],
+)