+++ mltu/tensorflow/model_utils.py
@@ -1,10 +1,8 @@
+import keras
 import typing
 import tensorflow as tf
-from tensorflow import keras
-from keras import layers
-from keras.models import Model
 
-class CustomModel(Model):
+class CustomModel(keras.models.Model):
     """ Custom TensorFlow model for debugging training process purposes
     """
     def train_step(self, train_data):
@@ -17,7 +15,7 @@ class CustomModel(Model):
             gradients = tape.gradient(loss, self.trainable_weights)
 
         # Applying the gradients on the model using the specified optimizer
-        self.optimizer.apply_gradients(zip(gradients, self.trainable_weights))
+        self.optimizer.apply_gradients(zip(gradients, self.trainable_weights)) # pyright: ignore
 
         # Update the metrics.
         # Metrics are configured in `compile()`.
@@ -41,19 +39,19 @@ class CustomModel(Model):
         return {m.name: m.result() for m in self.metrics}
 
 
-def activation_layer(layer, activation: str="relu", alpha: float=0.1) -> tf.Tensor:
+def activation_layer(layer, activation: str="relu", negative_slope: float=0.1) -> keras.layers.Layer:
     """ Activation layer wrapper for LeakyReLU and ReLU activation functions
     Args:
         layer: tf.Tensor
         activation: str, activation function name (default: 'relu')
-        alpha: float (LeakyReLU activation function parameter)
+        negative_slope: float (LeakyReLU activation function parameter)
     Returns:
         tf.Tensor
     """
     if activation == "relu":
-        layer = layers.ReLU()(layer)
+        layer = keras.layers.ReLU()(layer)
     elif activation == "leaky_relu":
-        layer = layers.LeakyReLU(alpha=alpha)(layer)
+        layer = keras.layers.LeakyReLU(negative_slope=negative_slope)(layer)
 
     return layer
 
@@ -67,29 +65,30 @@ def residual_block(
         padding: str = "same",
         kernel_initializer: str = "he_uniform",
         activation: str = "relu",
-        dropout: float = 0.2):
+        dropout: float = 0.2
+    ) -> keras.layers.Layer:
     # Create skip connection tensor
     x_skip = x
 
     # Perform 1-st convolution
-    x = layers.Conv2D(filter_num, kernel_size, padding = padding, strides = strides, kernel_initializer=kernel_initializer)(x)
-    x = layers.BatchNormalization()(x)
+    x = keras.layers.Conv2D(filter_num, kernel_size, padding = padding, strides = strides, kernel_initializer=kernel_initializer)(x)
+    x = keras.layers.BatchNormalization()(x)
     x = activation_layer(x, activation=activation)
 
     # Perform 2-nd convoluti
-    x = layers.Conv2D(filter_num, kernel_size, padding = padding, kernel_initializer=kernel_initializer)(x)
-    x = layers.BatchNormalization()(x)
+    x = keras.layers.Conv2D(filter_num, kernel_size, padding = padding, kernel_initializer=kernel_initializer)(x)
+    x = keras.layers.BatchNormalization()(x)
 
     # Perform 3-rd convolution if skip_conv is True, matchin the number of filters and the shape of the skip connection tensor
     if skip_conv:
-        x_skip = layers.Conv2D(filter_num, 1, padding = padding, strides = strides, kernel_initializer=kernel_initializer)(x_skip)
+        x_skip = keras.layers.Conv2D(filter_num, 1, padding = padding, strides = strides, kernel_initializer=kernel_initializer)(x_skip)
 
     # Add x and skip connection and apply activation function
-    x = layers.Add()([x, x_skip])     
+    x = keras.layers.Add()([x, x_skip])     
     x = activation_layer(x, activation=activation)
 
     # Apply dropout
     if dropout:
-        x = layers.Dropout(dropout)(x)
+        x = keras.layers.Dropout(dropout)(x)
 
     return x